+++ src/main/kotlin/cloud/glitchdev/rfu/config/seacreatures/SeaCreatureSettings.kt
@@ -9,5 +9,30 @@ data class SeaCreatureSettings(
 }
 
 data class SeaCreatureSetting(
-    val name: String?, val plural: String?, val article: String?, val special: Boolean?, val format: String?
+    val name: String?,
+    val plural: String?,
+    val article: String?,
+    val special: Boolean?,
+    val style: String?,
+    val catchMessage: String?,
+    val liquidType: String?,
+    val category: String?,
+    val lsRangeExcluded: Boolean?,
+    val conditions: SeaCreatureConditions?,
+    val bossbar: Boolean?
+)
+
+data class SeaCreatureConditions(
+    val isFestival: Boolean? = null,
+    val isSpooky: Boolean? = null,
+    val hotspot: String? = null,
+    val bait: String? = null,
+    val coords: List<SeaCreatureCoordRange>? = null
+)
+
+data class SeaCreatureCoordRange(
+    val x0: Double, val x1: Double,
+    val y0: Double, val y1: Double,
+    val z0: Double, val z1: Double,
+    val exclude: Boolean? = false
 )